requirements.txt
# Core / UI
streamlit==1.35.0
packaging==23.2
pillow==10.3.0

# Data / numerical
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
joblib==1.3.2
statsmodels==0.14.1

# ML
scikit-learn==1.3.2
imbalanced-learn==0.12.0
xgboost==2.0.3
lightgbm==4.3.0
catboost==1.2.5

# Deep learning
tensorflow==2.16.2          # Keras intégré
# PyTorch (CPU/CUDA as needed; use official install command for cuda)
torch==2.2.2
torchvision==0.17.2
pytorch-msssim

# Vision / augmentation
albumentations==1.4.3
albucore==0.0.11

# Profiling / reporting
ydata-profiling==4.7.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
kaleido==0.2.1

# MLOps / tracking
mlflow==2.8.1
python-dotenv==0.21.0
pydantic==2.12.0

# Extras / utils
openpyxl==3.1.2
pyarrow==14.0.2
fastparquet==2024.2.0
psutil==5.9.8
reportlab==4.1.0
psycopg2-binary==2.9.9

# --- Dask complet pour calcul distribué ---
dask
dask[dataframe]

# Dev / test
pytest==7.4.2
ruff==0.3.4

